Improved the accuracy of percentages in the pie charts to less than  “one percent margin of error”.

+++ src/EarningsTracker.Services/Revenues/IRevenueData.cs
@@ -6,7 +6,7 @@
         // name of a specific office
         string Name { get; set; }
         // total revenue percent for a specific office
-        int Percent { get; set; }
+        decimal Percent { get; set; }
         // total revenue amount for a specific office
         decimal Revenue { get; set; }
         // total revenue amount for other offices
